refactor(web): extract service tier for /api/v1/search out of 'use server' action
Public REST handlers under apps/web/app/api/v1/* were importing and
calling server-action exports as if they were plain service functions.
A "use server" file in Next.js is not just a "server-only module" — every
exported function becomes an RPC endpoint with its own auto-generated
POST URL, its own serialization boundary, its own per-call security ID,
and a requirement that arguments be serializable.

This commit pulls the implementation of `searchJobs`, `listTopCompanies`,
`listTopCompaniesAnonymous`, `getPostingDetail`, `getCurrencyRates`,
`getSalaryHistogram`, `getExperienceHistogram`, `loadMorePostings`, and
`parseSearchFilters` out of `apps/web/src/lib/actions/` into a new
`apps/web/src/lib/services/` tier that is server-only (via
`import "server-only"`) but NOT a "use server" module.

- `lib/services/search.ts` — full implementation, no "use server"
- `lib/services/search-input.ts` — full implementation, no "use server"
- `lib/actions/search.ts` — thin "use server" wrapper re-exporting from
  `lib/services/search` so existing UI callers (search-bar, search-page,
  company-page, explore-data, etc.) keep importing the same module path
  with no churn
- `lib/actions/search-input.ts` — same pattern for `parseSearchFilters`
- `app/api/v1/search/route.ts`, `app/api/v1/job/route.ts`,
  `app/api/v1/watchlist/create/route.ts` — import directly from
  `@/lib/services/*` instead of `@/lib/actions/*`
- `app/api/v1/search/route.test.ts` — mock the service modules to match
  the route's real import graph
- new `lib/services/__tests__/search-tier.test.ts` — pin the boundary:
  service files have no "use server", action wrappers DO have
  "use server", callable surfaces are identical, and `parseSearchFilters`
  is reachable as a plain async function from the service tier

Scope is intentionally narrow per #3231's "concrete first slice"
recommendation — only the search.ts + search-input.ts modules
(highest-impact, called by 3 of the 8 v1 routes). Follow-up issues will
file the same pattern for locations.ts, taxonomy.ts, company.ts, and
watchlists.ts.

Closes #3231 partial.
